# Propagate array dimensions from allocations