--- a/src/llvm-alloc-helpers.cpp
+++ b/src/llvm-alloc-helpers.cpp
@@ -302,3 +302,55 @@ void jl_alloc::runEscapeAnalysis(llvm::Instruction *I, EscapeAnalysisRequiredArg
         }
     }
 }
+
+//FIXME: This doesn't actually work on Windows, as Windows inserts a trampoline
+//function which doesn't have the same function pointer as the bare alloc_array
+//functions
+bool jl_alloc::getArrayAllocInfo(AllocIdInfo &info, llvm::CallInst *call) {
+    if (auto cexpr = dyn_cast<ConstantExpr>(call->getCalledOperand())) {
+        if (cexpr->getOpcode() == Instruction::IntToPtr) {
+            if (auto cint = dyn_cast<ConstantInt>(cexpr->getOperand(0))) {
+                std::size_t faddr = cint->getZExtValue();
+                if (faddr == reinterpret_cast<std::uintptr_t>(jl_alloc_array_1d)) {
+                    assert(call->getNumArgOperands() == 2);
+                    info.array.dimcount = 1;
+                } else if (faddr == reinterpret_cast<std::uintptr_t>(jl_alloc_array_2d)) {
+                    assert(call->getNumArgOperands() == 3);
+                    info.array.dimcount = 2;
+                } else if (faddr == reinterpret_cast<std::uintptr_t>(jl_alloc_array_3d)) {
+                    assert(call->getNumArgOperands() == 4);
+                    info.array.dimcount = 3;
+                } else if (faddr == reinterpret_cast<std::uintptr_t>(jl_new_array)) {
+                    assert(call->getNumArgOperands() == 2);
+                    info.array.dimcount = 0;
+                } else {
+                    return false;
+                }
+                info.isarray = true;
+                info.type = call->getArgOperand(0);
+                return true;
+            }
+        }
+    }
+    return false;
+}
+
+bool jl_alloc::getAllocIdInfo(AllocIdInfo &info, llvm::CallInst *call, llvm::Function *alloc_obj_func) {
+    auto callee = call->getCalledOperand();
+    if (callee == alloc_obj_func) {
+        info.isarray = false;
+        auto sz = call->getArgOperand(1);
+        info.type = call->getArgOperand(2);
+        if (auto size = dyn_cast<ConstantInt>(sz)) {
+            info.object.size = size->getZExtValue();
+            if (info.object.size < IntegerType::MAX_INT_BITS / 8 && info.object.size < INT32_MAX) {
+                return true;
+            }
+        }
+        info.object.size = -1;
+        return true;
+    } else {
+        return getArrayAllocInfo(info, call);
+    }
+    return false;
+}