… src/llvm-arraylen-opt.cpp
… test/llvmpasses/array-opt.ll