2D and 3D array dimension lengths cannot change, and thus we can replace loads of array length and array dimension size with their values from when the array was allocated. 1D array dimensions might change, but we can determine if they might using escape analysis. 

~~Currently, the mechanism for identifying array allocations does not work on Windows, due to the insertion of a trampoline function on Windows, which prevents identification by function pointer equivalence from working as intended (this data was collected from #43107). Fixing this for Windows would require overhauling how we handle ccalls during codegen, which is not part of this PR.~~

Identification of arrays is now done through metadata set during codegen, which should theoretically get around the Windows issue.

Depends on #43057 for creating the llvm-alloc-helpers.cpp file

Performance benchmark:

Function:
```
function f()
           s = 0
           a = zeros(Int, 100, 100, 100)
           for i in eachindex(a)
               s += a[i]
           end
           s
       end
```

Master:
Godbolt: https://godbolt.org/z/cj7h3e3Gr
Benchmark:
```
julia> @benchmark f()
BenchmarkTools.Trial: 3179 samples with 1 evaluation.
 Range (min … max):  917.762 μs …  13.191 ms  ┊ GC (min … max):  0.00% … 92.57%
 Time  (median):       1.413 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):     1.566 ms ± 802.684 μs  ┊ GC (mean ± σ):  10.83% ± 15.99%

  ▅▆█▄        ▂▇▆▃                  ▂▂▃▂  ▃▂    ▁▁  ▂   ▁▃▂   ▂ ▁
  ████▅▅▇█▄▁▁▁████▅▄▆▆▅▁▁▃▄▄▃▁▁▁▁▁▃█████▇███▇▄▄███▇███▅▆███▇▄██ █
  918 μs        Histogram: log(frequency) by time       3.24 ms <

 Memory estimate: 7.63 MiB, allocs estimate: 2.
```

PR:
Godbolt: https://godbolt.org/z/zKWz8dxev
Benchmark:
```
julia> @benchmark f()
BenchmarkTools.Trial: 6350 samples with 1 evaluation.
 Range (min … max):  687.484 μs …   4.911 ms  ┊ GC (min … max): 0.00% …  0.00%
 Time  (median):     702.864 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   782.345 μs ± 188.419 μs  ┊ GC (mean ± σ):  9.65% ± 14.78%

   █
  ▅█▇▃▄▂▃▃▂▂▂▂▂▂▂▂▁▁▂▁▁▁▁▁▁▁▂▂▁▁▁▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▂▃▄▃▃▂▂▂ ▂
  687 μs           Histogram: frequency by time         1.17 ms <

 Memory estimate: 7.63 MiB, allocs estimate: 2.
```