@pchintalapudi 2022-01-07
Compile time benchmarks:
---
Master:
```
Base  ───────────── 23.796904 seconds
ArgTools  ─────────  4.456852 seconds
Artifacts  ────────  0.102586 seconds
Base64  ───────────  0.105645 seconds
CRC32c  ───────────  0.008005 seconds
FileWatching  ─────  0.108983 seconds
Libdl  ────────────  0.001518 seconds
Logging  ──────────  0.036208 seconds
Mmap  ─────────────  0.085615 seconds
NetworkOptions  ───  0.117357 seconds
SHA  ──────────────  0.181987 seconds
Serialization  ────  0.270087 seconds
Sockets  ──────────  0.321430 seconds
Unicode  ──────────  0.077732 seconds
DelimitedFiles  ───  0.104025 seconds
LinearAlgebra  ────  7.762580 seconds
Markdown  ─────────  0.754207 seconds
Printf  ───────────  0.168311 seconds
Random  ───────────  1.242392 seconds
Tar  ──────────────  0.282602 seconds
Dates  ────────────  1.536529 seconds
Distributed  ──────  0.764613 seconds
Future  ───────────  0.005194 seconds
InteractiveUtils  ─  0.500062 seconds
LibGit2  ──────────  1.366151 seconds
Profile  ──────────  0.257939 seconds
SparseArrays  ─────  2.678179 seconds
UUIDs  ────────────  0.015961 seconds
REPL  ─────────────  3.456523 seconds
SharedArrays  ─────  0.500804 seconds
Statistics  ───────  0.188637 seconds
SuiteSparse  ──────  1.582334 seconds
TOML  ─────────────  0.067595 seconds
Test  ─────────────  0.315075 seconds
LibCURL  ──────────  0.438543 seconds
Downloads  ────────  0.335553 seconds
Pkg  ──────────────  4.502768 seconds
LazyArtifacts  ────  0.001633 seconds
Stdlibs total  ──── 34.712985 seconds
Sysimage built. Summary:
Total ───────  58.511190 seconds
Base: ───────  23.796904 seconds 40.6707%
Stdlibs: ────  34.712985 seconds 59.3271%
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1308/1339
Precompilation complete. Summary:
Total ─────── 109.996850 seconds
Generation ──  82.400907 seconds 74.9121%
Execution ───  27.595943 seconds 25.0879%
```
PR:
```
Base  ───────────── 23.928553 seconds
ArgTools  ─────────  4.422442 seconds
Artifacts  ────────  0.103806 seconds
Base64  ───────────  0.106303 seconds
CRC32c  ───────────  0.008324 seconds
FileWatching  ─────  0.110224 seconds
Libdl  ────────────  0.001559 seconds
Logging  ──────────  0.036711 seconds
Mmap  ─────────────  0.086969 seconds
NetworkOptions  ───  0.117836 seconds
SHA  ──────────────  0.184373 seconds
Serialization  ────  0.271228 seconds
Sockets  ──────────  0.322758 seconds
Unicode  ──────────  0.078247 seconds
DelimitedFiles  ───  0.105199 seconds
LinearAlgebra  ────  7.820788 seconds
Markdown  ─────────  0.758388 seconds
Printf  ───────────  0.151605 seconds
Random  ───────────  1.259426 seconds
Tar  ──────────────  0.285620 seconds
Dates  ────────────  1.510449 seconds
Distributed  ──────  0.776262 seconds
Future  ───────────  0.005317 seconds
InteractiveUtils  ─  0.503558 seconds
LibGit2  ──────────  1.377215 seconds
Profile  ──────────  0.258898 seconds
SparseArrays  ─────  2.681444 seconds
UUIDs  ────────────  0.016312 seconds
REPL  ─────────────  3.456168 seconds
SharedArrays  ─────  0.501174 seconds
Statistics  ───────  0.191785 seconds
SuiteSparse  ──────  1.591476 seconds
TOML  ─────────────  0.068235 seconds
Test  ─────────────  0.320033 seconds
LibCURL  ──────────  0.441098 seconds
Downloads  ────────  0.336550 seconds
Pkg  ──────────────  4.571783 seconds
LazyArtifacts  ────  0.001639 seconds
Stdlibs total  ──── 34.851874 seconds
Sysimage built. Summary:
Total ───────  58.781741 seconds
Base: ───────  23.928553 seconds 40.7075%
Stdlibs: ────  34.851874 seconds 59.2903%
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1308/1339
Precompilation complete. Summary:
Total ─────── 109.750784 seconds
Generation ──  81.986599 seconds 74.7025%
Execution ───  27.764185 seconds 25.2975%
```