@pchintalapudi 2022-01-08
# Other notable examples of optimization improvement:
---
## Example 2: doc/src/manual/performance-tips.md:952-963 loopinc

### Master:
Godbolt: https://godbolt.org/z/oT9Wa4784

### PR:
Godbolt: https://godbolt.org/z/r1or75rzK

## Example 3: doc/src/manual/performance-tips.md:969-984 loopinc_prealloc

### Master:
Godbolt: https://godbolt.org/z/7Ed4xd3o9

### PR:
Godbolt: https://godbolt.org/z/G6adnc69f

---
Both examples show complete removal of the bounds check due to length propagation, and LLVM then recognizes that all of the loads from the array can be forwarded and eliminates them all. When combined with #43548 and its ability to remove arrays that are never loaded from, and if xinc(!) is not marked noinline, LLVM will either vectorize loopinc (resulting in a massive speedup) or outright precompute the sum in loopinc_prealloc (another O(n) -> O(1) transformation)