@nanosoldier 2022-01-09
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/43487#issuecomment-1008126016) has completed - possible new issues were detected. A full report can be found [here](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/1f3cb02_vs_acb6e16/report.html).