@nanosoldier 2022-01-10
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/43487#issuecomment-1008458298) has completed - possible new issues were detected. A full report can be found [here](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/c1c20a8_vs_f7bb391/report.html).