@nanosoldier 2022-01-11
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/43487#issuecomment-1009182101) has completed - possible new issues were detected. A full report can be found [here](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/12b9a1b_vs_7b62896/report.html).