@vchuravy 2022-01-11
@nanosoldier `runbenchmarks(!"scalar", vs=":master")`