@nanosoldier 2022-01-11
[Your benchmark job](https://github.com/JuliaLang/julia/pull/43487#issuecomment-1009530271) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/b92fd65_vs_a7e4f6d/report.md).