@vchuravy 2022-01-14
@nanosoldier `runbenchmarks(!"scalar", vs=":master")`