@nanosoldier 2022-01-14
[Your benchmark job](https://github.com/JuliaLang/julia/pull/43487#issuecomment-1012679734) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/19b2f65_vs_294b0df/report.md).