@nanosoldier 2022-01-14
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/43487#issuecomment-1012679938) has completed - possible new issues were detected. A full report can be found [here](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/cc79187_vs_294b0df/report.html).