@vtjnash 2022-01-14
I am not certain if we want to move forward with tricks like this. The goal for Array was to replace it with something simpler implemented in Julia, which would hopefully allow the compiler to "see" this, without needing to make it a special case again in codegen (we had some code like this many years ago).