@vchuravy 2022-01-14
>  The goal for Array was to replace it with something simpler implemented in Julia, which would hopefully allow the compiler to "see" this

I think both Prem and I agree with that notion. But it is also the case that the move from `Array` to `Buffer` hasn't happened yet (and we are adding things like ImmutableArrays). So it might be a case of the perfect is the enemy of the good.